extract MailData struct to separate file;
